Moves HttpMethod enum to common location
Moves the `HttpMethod` enum from a React-specific location to a more general `common` directory.
This change promotes code reuse and avoids duplication, as the enum is used in multiple parts of the application.
